requirements.txt
Django==3.2.17
Pillow==9.4.0
torch==1.10.2
torchvision==0.11.3
